+++ frontend/components/ProductHeaderRatings/style.js
@@ -5,16 +5,12 @@ const { variables, colors } = themeConfig;
 
 const content = css({
   ' > div': {
-    // Apply styling to the RatingStars component to render the ProductHeader nicely.
+    // Apply styling to the RatingStars component to render the ProductHeaderRatings nicely.
     padding: variables.gap.big,
     marginBottom: 0,
     backgroundColor: colors.light,
     borderTop: 'rgba(0, 0, 0, 0.05) 2px solid',
-  },
-  ' + div': {
-    // Hide the original ProductHeader.
-    display: 'none',
-  },
+  }
 });
 
 export default {