+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
+## [2.0.2] - 2019-04-
+- No header is rendered on both gmd and ios templates.
+- Grouped product add to cart logic (picker) doesn't apply to variant products.
+- Grouped product ratings is shown again on PDP page
+
+TODO:
+error while opening the child product
+
 ## [2.0.1] - 2019-04-04
 - Bugfix for Variant GMD products.